+++ CHANGELOG.md
@@ -8,7 +8,14 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 
 ### Added
 
+- Added `GetSamFields` trait, which provides a way of unifying generic SAM data
+  (like `SamData`, `SamDataView`, and custom application structs)
 - The `Float` trait has been made public for unifying `f32` and `f64`
+
+### Changed
+
+- `BamWriter` now accepts a generic record implementing `GetSamFields` instead
+  of only `SamData`
 
 ## [0.0.33] - 2026-09-17